create_case_split
